third_party/blink/web_tests/external/wpt/trusted-types/set-attributes-mutations-in-callback.tentative-expected.txt

This is a testharness.js-based test.
Found 48 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Attr.value works for elementNS=http://www.w3.org/1999/xhtml, element=DIV, attrName=onclick (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Attr.value works for elementNS=http://www.w3.org/2000/svg, element=g, attrName=ondblclick (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Attr.value works for elementNS=http://www.w3.org/1998/Math/MathML, element=mrow, attrName=onmousedown (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Attr.value works for elementNS=http://www.w3.org/1999/xhtml, element=IFRAME, attrName=srcdoc (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Attr.value works for elementNS=http://www.w3.org/1999/xhtml, element=SCRIPT, attrName=src (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Attr.value works for elementNS=http://www.w3.org/2000/svg, element=script, attrName=href (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Attr.value works for elementNS=http://www.w3.org/2000/svg, element=script, attrNS=http://www.w3.org/1999/xlink, attrName=href (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/1999/xhtml, element=DIV, attrName=onclick (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/2000/svg, element=g, attrName=ondblclick (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/1998/Math/MathML, element=mrow, attrName=onmousedown (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/1999/xhtml, element=IFRAME, attrName=srcdoc (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/1999/xhtml, element=SCRIPT, attrName=src (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/2000/svg, element=script, attrName=href (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/2000/svg, element=script, attrNS=http://www.w3.org/1999/xlink, attrName=href (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.textContent works for elementNS=http://www.w3.org/1999/xhtml, element=DIV, attrName=onclick (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.textContent works for elementNS=http://www.w3.org/2000/svg, element=g, attrName=ondblclick (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.textContent works for elementNS=http://www.w3.org/1998/Math/MathML, element=mrow, attrName=onmousedown (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.textContent works for elementNS=http://www.w3.org/1999/xhtml, element=IFRAME, attrName=srcdoc (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.textContent works for elementNS=http://www.w3.org/1999/xhtml, element=SCRIPT, attrName=src (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.textContent works for elementNS=http://www.w3.org/2000/svg, element=script, attrName=href (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Node.textContent works for elementNS=http://www.w3.org/2000/svg, element=script, attrNS=http://www.w3.org/1999/xlink, attrName=href (delete attribute)
  assert_equals: expected 2 but got 3
[FAIL] Element.setAttribute works for elementNS=http://www.w3.org/1999/xhtml, element=DIV, attrName=onclick (modify attribute)
  assert_equals: expected 1 but got 2
[FAIL] Element.setAttribute works for elementNS=http://www.w3.org/2000/svg, element=g, attrName=ondblclick (modify attribute)
  assert_equals: expected 1 but got 2
[FAIL] Element.setAttribute works for elementNS=http://www.w3.org/1998/Math/MathML, element=mrow, attrName=onmousedown (modify attribute)
  assert_equals: expected 1 but got 2
[FAIL] Element.setAttribute works for elementNS=http://www.w3.org/1999/xhtml, element=IFRAME, attrName=srcdoc (modify attribute)
  assert_equals: expected 1 but got 2
[FAIL] Element.setAttribute works for elementNS=http://www.w3.org/1999/xhtml, element=SCRIPT, attrName=src (modify attribute)
  assert_equals: expected 1 but got 2
[FAIL] Element.setAttribute works for elementNS=http://www.w3.org/2000/svg, element=script, attrName=href (modify attribute)
  assert_equals: expected 1 but got 2
[FAIL] Attr.value works for elementNS=http://www.w3.org/1999/xhtml, element=DIV, attrName=onclick (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Attr.value works for elementNS=http://www.w3.org/2000/svg, element=g, attrName=ondblclick (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Attr.value works for elementNS=http://www.w3.org/1998/Math/MathML, element=mrow, attrName=onmousedown (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Attr.value works for elementNS=http://www.w3.org/1999/xhtml, element=IFRAME, attrName=srcdoc (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Attr.value works for elementNS=http://www.w3.org/1999/xhtml, element=SCRIPT, attrName=src (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Attr.value works for elementNS=http://www.w3.org/2000/svg, element=script, attrName=href (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Attr.value works for elementNS=http://www.w3.org/2000/svg, element=script, attrNS=http://www.w3.org/1999/xlink, attrName=href (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/1999/xhtml, element=DIV, attrName=onclick (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/2000/svg, element=g, attrName=ondblclick (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/1998/Math/MathML, element=mrow, attrName=onmousedown (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/1999/xhtml, element=IFRAME, attrName=srcdoc (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/1999/xhtml, element=SCRIPT, attrName=src (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/2000/svg, element=script, attrName=href (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.nodeValue works for elementNS=http://www.w3.org/2000/svg, element=script, attrNS=http://www.w3.org/1999/xlink, attrName=href (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.textContent works for elementNS=http://www.w3.org/1999/xhtml, element=DIV, attrName=onclick (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.textContent works for elementNS=http://www.w3.org/2000/svg, element=g, attrName=ondblclick (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.textContent works for elementNS=http://www.w3.org/1998/Math/MathML, element=mrow, attrName=onmousedown (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.textContent works for elementNS=http://www.w3.org/1999/xhtml, element=IFRAME, attrName=srcdoc (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.textContent works for elementNS=http://www.w3.org/1999/xhtml, element=SCRIPT, attrName=src (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.textContent works for elementNS=http://www.w3.org/2000/svg, element=script, attrName=href (move attribute node to another element)
  assert_equals: expected 0 but got 1
[FAIL] Node.textContent works for elementNS=http://www.w3.org/2000/svg, element=script, attrNS=http://www.w3.org/1999/xlink, attrName=href (move attribute node to another element)
  assert_equals: expected 0 but got 1
Harness: the test ran to completion.
